delete fixing

+++ main.py
@@ -126,17 +126,19 @@ def Appointments():
     
 @app.route('/DeleteAppointment/<int:appointment_id>', methods=['DELETE'])
 def delete_appointment(appointment_id):
+    
+        if request.method == 'DELETE':
 
-        # Check if the appointment with the given ID exists
-        appointment = TimeSetter.query.filter_by(id=appointment_id).first()
-        if appointment:
-            # Delete the appointment and commit the changes
-            db.session.delete(appointment)
-            db.session.commit()
-            return jsonify({"message": "Appointment deleted successfully", "status_code": 200}), 200
-        else:
-            # Return a 404 status code and a message if the appointment is not found
-            return jsonify({"message": "Appointment not found", "status_code": 404}), 404
+            # Check if the appointment with the given ID exists
+            appointment = TimeSetter.query.filter_by(id=appointment_id).first()
+            if appointment:
+                # Delete the appointment and commit the changes
+                db.session.delete(appointment)
+                db.session.commit()
+                return jsonify({"message": "Appointment deleted successfully", "status_code": 200}), 200
+            else:
+                # Return a 404 status code and a message if the appointment is not found
+                return jsonify({"message": "Appointment not found", "status_code": 404}), 404
 
 @app.route('/clients', methods=['GET'])
 def clients():
